Set '-ms-overflow-style: scrollbar' on the scrollbar sizer element
The default overflow style on tablets (for example the Microsoft Surface) is"-ms-autohiding-scrollbar". That causes simplebar to miscalculate the sidebar size (width:0) and that causes the native scroll bar to remain visible.

You can easily see that by visiting to the simplebar homepage on Microsoft Surface, using the Edge browser.