fix: re-write using cross-compatible render function
This was based on master, do note that there was prior (failed) work at building a cross compatible version that was reverted and this had to be brought into this commit (conditional test-utils, swap-vue, etc)

---

Using a SFC to create cross-compatible component lead us to unwanted bundle results when building the library. Apparently the template compiler used was the one for vue3. Manually changing rollup plugin vue compiler option to explicitly use other vue-template-compiler version did not change the results and even if did, we would need to compile 3 different versions and somewhat do a postinstall script that point consumers to the right one. Not sure how to do that so went with another approach.

The alternative implemented was to convert the component into a pure render function, middleware to adjust arguments to vue2 and 3 signatures. This way there's no transformation when bundling, only minification and transpiling (babel) making output could work as expected even with a slight decrease on readability for maintainers.